Experiment with adding to all_control directly
This is easy to get working to add new UI widgets, but still doesn't
help for the logic of mapping our keybinds. We still need a way of
hijacking the callbacks created in OptionsScreenn:_BuildControls()
in order to both update the binding text AND telling MapControl()
and OnControlMapped().

Also-- I'm not sure how to progmatically set a vanilla control either.
We need to restore CONTROL_PRIMARY to the original state after our
hijacked keybind setting is complete. TheInputProxy:CancelMapping()
seems to be a thing? but it's only ever called in the gamepad codepath.